drugi

+++ src/main/java/pl/edu/pw/finalproject/service/CourseService.java
@@ -1,6 +1,7 @@
 package pl.edu.pw.finalproject.service;
 
 import pl.edu.pw.finalproject.entities.Course;
+import pl.edu.pw.finalproject.entities.User;
 
 import java.util.List;
 
@@ -8,11 +9,13 @@ public interface CourseService {
 
     public List<Course> findAll();
 
-    public Course findById(int theId);
+    public Course findById(long theId);
 
     public void save(Course theCourse);
 
     public void deleteById(long theId);
+
+
 
 
 }
